Move push stuff to service worker context
