+++ frontend/src/employee-mobile-frontend/common/service-worker.tsx
@@ -3,7 +3,21 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import * as Sentry from '@sentry/browser'
-import React, { createContext, useEffect, useState } from 'react'
+import { differenceInDays } from 'date-fns'
+import React, {
+  createContext,
+  useContext,
+  useEffect,
+  useMemo,
+  useState
+} from 'react'
+
+import HelsinkiDateTime from 'lib-common/helsinki-date-time'
+import { UUID } from 'lib-common/types'
+import { mockNow } from 'lib-common/utils/helpers'
+
+import { upsertPushSubscription } from '../auth/api'
+import { UserContext } from '../auth/state'
 
 interface ServiceWorkerState {
   registration: ServiceWorkerRegistration | undefined
@@ -19,8 +33,19 @@ export const ServiceWorkerContextProvider = React.memo(
   }: {
     children: JSX.Element
   }) {
+    const user = useContext(UserContext).user.getOrElse(undefined)
+
     const [registration, setRegistration] =
       useState<ServiceWorkerRegistration>()
+
+    const pushNotifications = useMemo(() => {
+      if (!user?.pushApplicationServerKey) return undefined
+      if (!registration || !('pushManager' in registration)) return undefined
+      return new PushNotifications(user.id, registration.pushManager, {
+        userVisibleOnly: true,
+        applicationServerKey: user.pushApplicationServerKey
+      })
+    }, [user?.pushApplicationServerKey, user?.id, registration])
 
     useEffect(() => {
       registerServiceWorker()
@@ -29,6 +54,12 @@ export const ServiceWorkerContextProvider = React.memo(
           Sentry.captureException(err)
         })
     }, [])
+
+    useEffect(() => {
+      if (pushNotifications) {
+        pushNotifications.enable().catch((err) => Sentry.captureException(err))
+      }
+    }, [pushNotifications])
 
     const value = { registration }
 
@@ -46,5 +77,48 @@ const registerServiceWorker = async () => {
     return await navigator.serviceWorker.ready
   } else {
     return undefined
+  }
+}
+
+export class PushNotifications {
+  constructor(
+    private device: UUID,
+    private pushManager: PushManager,
+    private options: PushSubscriptionOptionsInit
+  ) {}
+
+  async enable() {
+    const sub = await this.refreshSubscription()
+    const authSecret = sub?.getKey('auth')
+    const ecdhKey = sub?.getKey('p256dh')
+    if (sub && authSecret && ecdhKey) {
+      await upsertPushSubscription(this.device, {
+        endpoint: sub.endpoint,
+        expires: sub.expirationTime
+          ? HelsinkiDateTime.fromSystemTzDate(new Date(sub.expirationTime))
+          : null,
+        authSecret: Array.from(new Uint8Array(authSecret)),
+        ecdhKey: Array.from(new Uint8Array(ecdhKey))
+      })
+    }
+  }
+
+  private async refreshSubscription(): Promise<PushSubscription | undefined> {
+    const state = await this.pushManager.permissionState(this.options)
+    if (state !== 'granted' && state !== 'prompt') {
+      return undefined
+    }
+    const sub = await this.pushManager.getSubscription()
+    if (sub) {
+      const now = mockNow() ?? new Date()
+      const expiringSoon = sub.expirationTime
+        ? differenceInDays(now, sub.expirationTime) < 7
+        : false
+      if (!expiringSoon) {
+        return sub
+      }
+      await sub.unsubscribe()
+    }
+    return await this.pushManager.subscribe(this.options)
   }
 }
